ADASP-23: adding a TaskPage component so that index stays clean
