+++ src/components/App/styles.css
@@ -29,114 +29,3 @@
   height: 32px;
   margin-top: 8px;
 }
-
-.createTaskButton {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  padding: 0px 32px;
-
-  width: 466px;
-  height: 40px;
-  left: 487px;
-  top: 274px;
-
-  background: rgba(65, 88, 196, 0.1);
-  border-radius: 4px;
-  border: none;
-}
-
-#newTaskText {
-  /* Button Text */
-
-  position: static;
-  height: 18px;
-  left: 35.09%;
-  right: 35.09%;
-  top: calc(50% - 18px / 2);
-
-  /* Button/Large */
-
-  font-style: normal;
-  font-weight: 600;
-  font-size: 18px;
-  line-height: 100%;
-  font-family: "Roboto";
-  /* identical to box height, or 18px */
-
-  text-align: center;
-  letter-spacing: 0.0125em;
-  text-transform: capitalize;
-
-  /* Primary */
-
-  color: #4158c4;
-
-  /* Inside Auto Layout */
-
-  flex: none;
-  order: 0;
-  flex-grow: 1;
-  margin: 0px 10px;
-}
-
-
-
-main {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-h5 {
-  font-family: "Roboto";
-  font-style: normal;
-  font-weight: bold;
-  font-size: 24px;
-  line-height: 100%;
-  /* identical to box height, or 24px */
-  color: rgba(0, 0, 0, 0.87);
-  
-}
-
-p {
-  width: 92px;
-  height: 16px;
-  left: 487px;
-  top: 128px;
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 16px;
-  line-height: 100%;
-  /* identical to box height, or 16px */
-  letter-spacing: 0.005em;
-  /* Text/Black/Active */
-  color: rgba(0, 0, 0, 0.87);
-}
-
-.complete {
-  width: 9em
-}
-
-.completeLabel {
-  display: flex;
-  align-items: center;
-  width: 28em;
-  justify-content: space-between;
-}
-
-.expand {
-  display: flex;
-  align-items: center;
-}
-
-.emptyTaskList {
-  width: 18em;
-  display: flex;
-}
-
-.emptyTaskList p {
-  flex: 1;
-}
