+++ src/components/TaskPage/README.md
@@ -1,0 +1,20 @@
+# `<TaskPage />`
+
+DESCRIPTION_HERE
+
+## Directory Structure
+
+- `stories.tsx`: Component playground (`npm run test:playground`)
+- `test.tsx`: Component tests (`npm run test:unit`)
+- `index.tsx`: Component code
+- `README.md`: Component documentation (hey, that's me!)
+
+## Example
+
+![TaskPage](../../../.loki/reference/chrome_TaskPage_example.png)
+
+> Image location: [`.loki/reference/chrome_TaskPage_example.png`](../../../.loki/reference/chrome_TaskPage_example.png)
+> 
+> To regenerate: 
+> 1. `npm run test:playground` (skip if running)
+> 1. `npm run test:visual:update -- --storiesFilter="^TaskPage example\$"`
