+++ src/index.tsx
@@ -1,18 +1,13 @@
 import React from "react"
 import ReactDOM from "react-dom"
-import { BrowserRouter, Route } from "react-router-dom"
 import { App } from "./components/App"
 import "./index.css"
-import { Homepage } from "./components/Homepage"
 import reportWebVitals from "./reportWebVitals"
 import * as serviceWorker from "./serviceWorker"
 
 ReactDOM.render(
   <React.StrictMode>
-    <BrowserRouter>
-      <Route exact path="/" component={Homepage} />
-      <Route exact path="/todo" component={App} />
-    </BrowserRouter>
+    <App />
   </React.StrictMode>,
   document.getElementById("root"),
 )
